Chat unreadmessages counts as Long instead Integer for better compatibility on common db count types handling

+++ Chat/src/main/java/com/cesar/Chat/dto/ConversationViewDTO.java
@@ -16,5 +16,5 @@ public class ConversationViewDTO {
     private LocalDateTime createdAt;
     private ConversationRecipientDTO recipient;
     private LastMessageDTO lastMessage;
-    private Integer unreadMessagesCount;
+    private Long unreadMessagesCount;
 }
